requirements.txt
hydra-core
pyk4a
opencv-python
numpy
pathlib
dataclasses